# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Blinker API
  description: Documentation globale de l'API Blinker
  version: 1.0.0
servers:
  - url: "http://localhost:3011/"
  - url: "https://dev.blinker.eterny.fr/"
...
paths:
...
  /follows/{targetUserID}:
    post:
      tags:
        - Follow
      summary: Suivre un utilisateur
      description: Permet à un utilisateur authentifié de suivre un autre utilisateur.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: targetUserID
          required: true
          schema:
            type: string
          description: ID de l'utilisateur à suivre.
      responses:
        '200':
          description: L'utilisateur est maintenant suivi.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilisateur suivi"
        '400':
          description: Erreur de validation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Erreur de validation"
        '404':
          description: Utilisateur non trouvé.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilisateur non trouvé"
...
  /follows/followers/{userID}:
    get:
      tags:
        - Follow
      summary: Obtenir la liste des abonnés d'un utilisateur
      description: Renvoie la liste paginée des utilisateurs qui suivent un utilisateur donné, avec des informations simplifiées sur chaque abonné (avatar, nom et nom d'utilisateur).
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userID
          required: true
          schema:
            type: string
          description: ID de l'utilisateur cible.
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
          description: Numéro de la page à retourner.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
          description: Nombre d'éléments par page.
      responses:
        '200':
          description: Liste paginée des abonnés retournée avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 10
                  total:
                    type: integer
                    example: 42
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        avatar_url:
                          type: string
                          example: "https://example.com/avatar.jpg"
                        name:
                          type: string
                          example: "John Doe"
                        username:
                          type: string
                          example: "johndoe"
        '404':
          description: Utilisateur non trouvé.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilisateur non trouvé"
  /follows/following/{userID}:
    get:
      tags:
        - Follow
      summary: Obtenir la liste des utilisateurs suivis par un utilisateur
      description: Renvoie la liste paginée des utilisateurs suivis par un utilisateur donné, avec des informations simplifiées sur chaque utilisateur suivi (avatar, nom et nom d'utilisateur).
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userID
          required: true
          schema:
            type: string
          description: ID de l'utilisateur cible.
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
          description: Numéro de la page à retourner.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
          description: Nombre d'éléments par page.
      responses:
        '200':
          description: Liste paginée des utilisateurs suivis retournée avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 10
                  total:
                    type: integer
                    example: 35
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        avatar_url:
                          type: string
                          example: "https://example.com/avatar.jpg"
                        name:
                          type: string
                          example: "Jane Doe"
                        username:
                          type: string
                          example: "janedoe"
        '404':
          description: Utilisateur non trouvé.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilisateur non trouvé"

  # ---------------------------
  # Endpoints Authentification
  # ---------------------------
...
  /login:
    post:
      tags:
        - Authentification
      summary: Authentification via email/mot de passe
      description: Authentifie un utilisateur à l'aide de son email et de son mot de passe.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "test@test.com"
                password:
                  type: string
                  example: "test"
      responses:
        '200':
          description: Authentification réussie.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "jwt_token"
        '400':
          description: Erreur de validation ou d'authentification.
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
  /checkToken:
    post:
      tags:
        - Authentification
      summary: Vérifier un token JWT
      description: Vérifie la validité d'un token JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  example: "jwt_token"
      responses:
        '200':
          description: Token valide.
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
        '400':
          description: Token invalide.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token invalide"
  /login/google:
    post:
      tags:
        - Authentification
      summary: Authentification avec Google OAuth
      description: Authentifie un utilisateur via Google OAuth.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id_token
              properties:
                id_token:
                  type: string
                  example: "google_id_token"
      responses:
        '200':
          description: Authentification réussie.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "jwt_token"
        '400':
          description: Erreur d'authentification.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  # ---------------------------
  # Endpoints Profiles
  # ---------------------------
  /profiles:
    post:
      tags:
        - Profiles
      summary: Créer un profil pour un utilisateur
      description: Associe un profil à un utilisateur existant avec un username, une bio et une photo.
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Fichier image de l'avatar à uploader.
                userID:
                  type: string
                  format: uuid
                  example: "550e8400-e29b-41d4-a716-446655440000"
                username:
                  type: string
                  example: "user123"
                bio:
                  type: string
                  example: "Développeur full-stack passionné."
      responses:
        '201':
          description: Profil créé avec succès.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Erreur de validation.
        '409':
          description: Username déjà utilisé.
        '500':
          description: Erreur interne.
  /profiles/{userID}:
    get:
      tags:
        - Profiles
      summary: Récupérer le profil d'un utilisateur
      description: Retourne le profil d'un utilisateur à partir de son userID.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userID
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Profil récupéré avec succès.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '404':
          description: Profil non trouvé.
        '500':
          description: Erreur interne.
  /profiles/upload-avatar/{userID}:
    post:
      tags:
        - Profiles
      summary: Uploader un avatar pour un utilisateur
      description: Met à jour l'avatar d'un utilisateur en uploadant une nouvelle image.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userID
          schema:
            type: string
            format: uuid
          required: true
          description: UUID de l'utilisateur
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Fichier image de l'avatar à uploader.
      responses:
        '200':
          description: Avatar uploadé avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Avatar uploaded successfully"
                  avatarUrl:
                    type: string
                    example: "/uploads/avatar.png"
        '400':
          description: Erreur de validation ou fichier manquant.
        '404':
          description: Utilisateur non trouvé.
        '500':
          description: Erreur interne du serveur.

  # ---------------------------
  # Endpoints Users
  # ---------------------------
...
  /interactions/like/{postID}:
    post:
      tags:
        - Interactions
      summary: Liker un Blink
      description: Ajoute un like à un Blink ou annule un like existant.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: postID
          in: path
          required: true
          schema:
            type: string
          description: ID du Blink à liker.
      responses:
        '200':
          description: Like ajouté, supprimé ou mis à jour avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: boolean
                    example: true
                  removed:
                    type: boolean
                    example: false
                  updated:
                    type: boolean
                    example: false
        '401':
          description: Utilisateur non authentifié.
        '500':
          description: Erreur interne du serveur.
  /interactions/dislike/{postID}:
    post:
      tags:
        - Interactions
      summary: Disliker un Blink
      description: Ajoute un dislike à un Blink ou annule un dislike existant.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: postID
          in: path
          required: true
          schema:
            type: string
          description: ID du Blink à disliker.
      responses:
        '200':
          description: Dislike ajouté, supprimé ou mis à jour avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: boolean
                    example: true
                  removed:
                    type: boolean
                    example: false
                  updated:
                    type: boolean
                    example: false
        '401':
          description: Utilisateur non authentifié.
        '500':
          description: Erreur interne du serveur.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'JWT'
      in: 'header'
      name: 'Authorization'
  schemas:
...
    Profile:
      type: object
      properties:
        userID:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        display_name:
          type: string
          example: "John Doe"
        username:
          type: string
          example: "user123"
        bio:
          type: string
          example: "Développeur full-stack passionné."
        avatar_url:
          type: string
          example: "https://example.com/avatar.png"
        score:
          type: integer
          example: 0